feat: introduce Legacy trait for PASETO versions
This commit introduces a new `Legacy` trait to mark PASETO versions that do not support implicit assertions. The `Legacy` trait is applied to v1 and v2 tokens. Additionally, the code has been refactored to use this new trait in place of the previous `Modern` trait for these versions. The import statements have also been updated accordingly.